LayoutTests/crypto/workers/subtle/ec-import-jwk-key-private-expected.txt

[Worker] Test importing a JWK ECDH private key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ec-import-jwk-key-private.js
[Worker] Importing a key...
PASS [Worker] privateKey.toString() is '[object CryptoKey]'
PASS [Worker] privateKey.type is 'private'
PASS [Worker] privateKey.extractable is false
PASS [Worker] privateKey.algorithm.name is 'ECDH'
PASS [Worker] privateKey.algorithm.namedCurve is 'P-256'
PASS [Worker] privateKey.usages is ['deriveBits', 'deriveKey']
PASS successfullyParsed is true

TEST COMPLETE
